# perf: improve prod time-to-interactive by ~9% by enabling minifier name mangling