#### Description of changes

Enabling name mangling in our prod minifier settings (causes it to rename variables during minification to be single characters) decreases our prod bundle sizes by ~29% and decreased time to interactive on the details page (measured as "the point where the loading spinner goes away and the main page starts animating in) by ~9% (from 770ms +- 15ms to 710ms +- 20ms).

It's extremely rare for us to have to debug in a prod extension build (even in the worst case, if a customer needs to perform a repro on our behalf, they can do so with a canary/playground build if necessary to get debug bundles).

**without mangling (before this PR):**
```
                Asset      Size  Chunks             Chunk Names
 background.bundle.js  4.13 MiB       0  [emitted]  background
       background.css  5.96 KiB       0  [emitted]  background
detailsView.bundle.js  4.37 MiB       1  [emitted]  detailsView
      detailsView.css  11.4 KiB       1  [emitted]  detailsView
   devtools.bundle.js   161 KiB       2  [emitted]  devtools
   injected.bundle.js  4.11 MiB       3  [emitted]  injected
         injected.css  5.96 KiB       3  [emitted]  injected
   insights.bundle.js  3.05 MiB       4  [emitted]  insights
      popup.bundle.js  3.59 MiB       5  [emitted]  popup
            popup.css  5.96 KiB       5  [emitted]  popup
```

**with mangling (after this PR):**
```
                Asset      Size  Chunks             Chunk Names
 background.bundle.js  2.94 MiB       0  [emitted]  background
       background.css  5.96 KiB       0  [emitted]  background
detailsView.bundle.js  3.15 MiB       1  [emitted]  detailsView
      detailsView.css  11.4 KiB       1  [emitted]  detailsView
   devtools.bundle.js  85.2 KiB       2  [emitted]  devtools
   injected.bundle.js  2.92 MiB       3  [emitted]  injected
         injected.css  5.96 KiB       3  [emitted]  injected
   insights.bundle.js  2.16 MiB       4  [emitted]  insights
      popup.bundle.js  2.51 MiB       5  [emitted]  popup
            popup.css  5.96 KiB       5  [emitted]  popup
```
#### Pull request checklist

- [n/a] Addresses an existing issue: Fixes #0000
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS